Add 'reset' functionality to CODAP API
It would be useful if the table structure changes during the game. It resets the document thus allowing creating new collections and adding new cases in them.

TODO: Currently, the graph and table views should be closed before 'reset' is used since the browser's table structure is currently not refreshed on change of table/graph after the creation of a new document.